+++ packages/odyssey-react-mui/src/Autocomplete.tsx
@@ -18,6 +18,7 @@ import {
 import { memo, useCallback } from "react";
 
 import { Field } from "./Field";
+import { FieldComponentProps } from "./FieldComponentProps";
 import type { SeleniumProps } from "./SeleniumProps";
 
 export type AutocompleteProps<
@@ -25,10 +26,6 @@ export type AutocompleteProps<
   HasMultipleChoices extends boolean | undefined,
   IsCustomValueAllowed extends boolean | undefined
 > = {
-  /**
-   * The error message for the Select
-   */
-  errorMessage?: string;
   /**
    * Enables multiple choice selection
    */
@@ -38,14 +35,6 @@ export type AutocompleteProps<
     undefined,
     IsCustomValueAllowed
   >["multiple"];
-  /**
-   * The hint text for the Autocomplete input
-   */
-  hint?: string;
-  /**
-   * The id attribute of the Select
-   */
-  id?: string;
   /**
    * Allows the input of custom values
    */
@@ -74,10 +63,6 @@ export type AutocompleteProps<
     IsCustomValueAllowed
   >["loading"];
   /**
-   * If `true`, the `input` element is not required.
-   */
-  isOptional?: boolean;
-  /**
    * Makes the Autocomplete input read-only
    */
   isReadOnly?: MuiAutocompleteProps<
@@ -90,10 +75,6 @@ export type AutocompleteProps<
    * The label text for the autocomplete input
    */
   label: string;
-  /**
-   * The name of the `input` element. Defaults to the `id` if not set.
-   */
-  name?: string;
   /**
    * Callback fired when the autocomplete loses focus.
    */
@@ -148,7 +129,11 @@ export type AutocompleteProps<
     undefined,
     IsCustomValueAllowed
   >["value"];
-} & SeleniumProps;
+} & Pick<
+  FieldComponentProps,
+  "errorMessage" | "hint" | "id" | "isOptional" | "name"
+> &
+  SeleniumProps;
 
 const Autocomplete = <
   OptionType,